requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
spacy==3.8.7
sqlalchemy==2.0.44
pymysql==1.1.2
python-dotenv==1.0.1
requests==2.32.5
pytesseract==0.3.13
python-multipart==0.0.20
httpx==0.28.1
Pillow==12.0.0
greenlet==3.2.4
annotated-types==0.7.0
typing-extensions==4.15.0
anyio==4.11.0
